# tree: reduce "rebasing commits" benchmarks when not in perf mode